Moved the pre-flight checks from cli to cluster crate
