+++ src/cluster/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/lib.rs"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+colored = "2.0.0"
 semver = "0.10.0"
 url = "2.1.1"
 tracing = "0.1.19"
@@ -22,6 +23,8 @@ tracing-futures = "0.2.4"
 serde = "1.0.115"
 serde_json = "1.0.57"
 thiserror = "1.0.20"
+async-trait = "0.1.21"
+
 
 # Fluvio dependencies
 fluvio = { version = "0.2.0", path = "../client" }
